silencing implicit cast warnings

+++ MaVec-Demo/View Controllers/LeastSquaresApproximationViewController.m
@@ -50,11 +50,11 @@
     self.navigationController.navigationBar.translucent = NO;
     self.approximationView.order = 3;
     self.orderLabel.text = @"2";
-    [self.orderSlider setValue:0.3 animated:YES];
+    [self.orderSlider setValue:0.3f animated:YES];
 }
 
 - (IBAction)changedOrder:(id)sender {
-    int order = self.orderSlider.value * 10;
+    int order = (int)self.orderSlider.value * 10;
     self.approximationView.order = order;
     self.orderLabel.text = [NSString stringWithFormat:@"%u", order-1];
 }
